feat: persons page suspense + some table suspense fixes
