+++ apps/web/src/app/(dashboard)/(account)/profiel/[handle]/_components/logbook/logbook-table.tsx
@@ -201,7 +201,10 @@ export function LogbookTable({
         <SortableTableHead table={table} />
         <TableBody>
           {placeholderRows && placeholderRows > 0 ? (
-            <PlaceholderTableRows rows={placeholderRows} />
+            <PlaceholderTableRows
+              rows={placeholderRows}
+              colSpan={columns.length}
+            />
           ) : (
             <>
               <NoTableRows table={table}>
