+++ apps/web/src/app/(dashboard)/(management)/locatie/[location]/diplomas/_components/table.tsx
@@ -153,7 +153,10 @@ export default function CertificateTable({
         <DefaultTableHead table={table} />
         <TableBody>
           {placeholderRows && placeholderRows > 0 ? (
-            <PlaceholderTableRows rows={placeholderRows} />
+            <PlaceholderTableRows
+              rows={placeholderRows}
+              colSpan={columns.length}
+            />
           ) : (
             <>
               <NoTableRows table={table}>Geen items gevonden</NoTableRows>
