+++ apps/web/src/app/(dashboard)/_components/table-content.tsx
@@ -75,21 +75,24 @@ export function DefaultTableCell<T>({
 
 export function PlaceholderTableRows({
   rows,
+  colSpan,
 }: {
   rows: number;
+  colSpan: number;
 }) {
   return [...Array(rows)].map((_, index) => (
     <tr
       // biome-ignore lint/suspicious/noArrayIndexKey: no other option
       key={`logbook-table-placeholder-${index}`}
-      className="relative h-11.25"
     >
-      <span
-        className="inline-block top-0 left-0 absolute bg-slate-200 mt-2 rounded w-full h-9.25 animate-pulse"
-        style={{
-          animationDelay: `${index * 0.3}s`,
-        }}
-      />
+      <td className="w-full h-11.25" colSpan={colSpan}>
+        <span
+          className="inline-block bg-slate-200 mt-2 rounded w-full h-9.25 animate-pulse"
+          style={{
+            animationDelay: `${index * 0.3}s`,
+          }}
+        />
+      </td>
     </tr>
   ));
 }
